requirements.txt
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
boto==2.49.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
confusable-homoglyphs==3.3.1
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==43.0.0
dbfread==2.0.7
Deprecated==1.2.14
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-filter==24.3
django-formtools==2.5.1
django-lockdown==4.0.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-yasg==1.21.7
dulwich==0.22.1
et-xmlfile==1.1.0
fastjsonschema==2.20.0
filelock==3.15.4
googleapis-common-protos==1.64.0
greenlet==3.0.3
grpcio==1.66.0
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.0.0
inflection==0.5.1
installer==0.7.0
isodate==0.6.1
itypes==1.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.3.0
leather==0.4.0
Markdown==3.7
MarkupSafe==2.1.5
more-itertools==10.4.0
msgpack==1.1.0rc2
numpy==2.1.0
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
packaging==24.1
parsedatetime==2.6
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
protobuf==4.25.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
str2bool==1.1
text-unidecode==1.3
tomlkit==0.13.2
trove-classifiers==2024.7.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
wheel==0.44.0
whitenoise==6.7.0
wrapt==1.16.0
xlrd==2.0.1
zipp==3.20.1
